# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Test Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/terms": {
            "get": {
                "tags": [
                    "Terms"
                ],
                "summary": "Get list of terms with pagination",
                "operationId": "e3879f0f8d4c89ae63a5ab3c7201cbaf",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Response message",
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "total": {
                                            "description": "Total number of terms",
                                            "type": "integer"
                                        },
                                        "terms": {
                                            "description": "List of terms",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "version": {
                                                        "description": "Version of the term",
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "description": "Title of the term",
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "description": "Content of the term",
                                                        "type": "string"
                                                    },
                                                    "apply_date": {
                                                        "description": "Apply date of the term",
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "description": "Current page",
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "description": "Last page",
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "description": "Number of items per page",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Terms"
                ],
                "summary": "Create a new term",
                "description": "Creates a new term with the provided details.",
                "operationId": "storeTerm",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "version",
                                    "title",
                                    "apply_date"
                                ],
                                "properties": {
                                    "version": {
                                        "type": "string",
                                        "example": "1.0.0"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Terms Title"
                                    },
                                    "apply_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-20"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Optional content",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Term created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0.0"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Terms Title"
                                        },
                                        "apply_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-03-20"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Optional content",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "version": [
                                                    "The version field is required."
                                                ],
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Bearer token in this format: `Bearer {token}`",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}